delete attribute module
