+++ lib/musical_score/element.rb
@@ -3,6 +3,8 @@ require "musical_score/element/measure"
 require "musical_score/element/note"
 require "musical_score/element/pitch"
 require "musical_score/element/const"
+require "musical_score/element/clef"
+require "musical_score/element/time"
 
 module MusicalScore
     module Element
