+++ lib/musical_score/element/attribute.rb
@@ -1,9 +1,6 @@
-require "musical_score/element/attribute/clef"
-require "musical_score/element/attribute/time"
-
 module MusicalScore
     module Element
-        module Attribute
+        class Attribute
         end
     end
 end
