--- lib/musical_score/element/attribute/time.rb
@@ -1,22 +1,0 @@
-module MusicalScore
-    module Element
-        module Attribute
-            class Time
-                attr_reader :beats, :beat_type
-
-                # constructor
-                #
-                # @param beats [Integer]
-                # @param beat_type [Integer]
-                def initialize(beats, beat_type)
-                    @beats     = beats
-                    @beat_type = beat_type
-                end
-
-                def to_s
-                    return "%d / %d" % [@beats, @beat_type]
-                end
-            end
-        end
-    end
-end
