+++ spec/element/clef_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe MusicalScore::Element::Attribute::Clef do
+describe MusicalScore::Element::Clef do
     describe 'initialize' do
-        let(:clef) { MusicalScore::Element::Attribute::Clef.new(:G)}
+        let(:clef) { MusicalScore::Element::Clef.new(:G)}
         it 'success case' do
             expect(clef.sign).to eq :G
             expect(clef.line).to eq 0
@@ -10,7 +10,7 @@ describe MusicalScore::Element::Attribute::Clef do
         end
 
         it 'raise error' do
-            expect{ MusicalScore::Element::Attribute::Clef.new(:D)}.to raise_error(MusicalScore::InvalidClefSign)
+            expect{ MusicalScore::Element::Clef.new(:D)}.to raise_error(MusicalScore::InvalidClefSign)
         end
     end
 end
